lab06/requirements.txt
scapy
psutil
python-telegram-bot